projects/21-nimbus-3000/requirements.txt
# Core ML/DL libraries
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
pandas==2.1.1
scikit-learn==1.3.1

# Optimizers (Adan requires external package)
adan-pytorch==0.1.0

# Image processing
pillow==10.1.0
opencv-python==4.8.1.78

# Visualization
matplotlib==3.8.0
seaborn==0.13.0

# Progress bars
tqdm==4.66.1

# Utilities
pyyaml==6.0.1

# Logging (optional)
wandb==0.15.12

# Statistical analysis
scipy==1.11.3

# Report generation
reportlab==4.0.7
